requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.2
requests==2.28.2
pandas==2.0.0
openpyxl==3.1.2
camelot-py==0.11.0
